requirements.txt
Flask == 1.1.2
Pillow == 8.4.0
numpy == 1.20.3
pandas == 1.3.4
wordcloud == 1.8.1
